+++ src/ast/semantic_format.rs
@@ -1793,6 +1793,9 @@ impl<'document, 'labels, 'writer, W: io::Write + ?Sized>
                         super::PartUsageBodyElement::Annotation(_member) => {
                             self.write_marker(&mut first, "annotation")?;
                         }
+                        super::PartUsageBodyElement::InOutDecl(_declaration) => {
+                            self.write_marker(&mut first, "in-out-declaration")?;
+                        }
                         super::PartUsageBodyElement::AttributeUsage(usage) => {
                             self.write_item_prefix(&mut first)?;
                             self.write_attribute_usage(&usage.value)?;